@@ -1,28 +1,32 @@
-import React from "react";
-import { Switch, Route, Router } from "react-router-dom";
-import {
-  StylesProvider,
-  createGenerateClassName,
-} from "@material-ui/core/styles";
+import React, { useState, useEffect } from 'react';
+import { Routes, Route, Router } from 'react-router-dom';
+import { StyledEngineProvider } from '@mui/material/styles';
 
-import Signin from "./components/Signin";
-import Signup from "./components/Signup";
+import Signin from './components/Signin';
+import Signup from './components/Signup';
 
-const generateClassName = createGenerateClassName({
-  productionPrefix: "au",
-});
+export default ({ history, onSignIn }) => {
+  const [location, setLocation] = useState(history.location);
 
-export default ({ history }) => {
+  useEffect(() => {
+    // Listen to history changes and update location state
+    const unlisten = history.listen((update) => {
+      setLocation(update.location);
+    });
+
+    return unlisten; // Clean up the listener on unmount
+  }, [history]);
+
   return (
     <div>
-      <StylesProvider generateClassName={generateClassName}>
-        <Router history={history}>
-          <Switch>
-            <Route path="/auth/signin" component={Signin} />
-            <Route path="/auth/signup" component={Signup} />
-          </Switch>
+      <StyledEngineProvider injectFirst>
+        <Router location={location} navigator={history}>
+          <Routes>
+            <Route path="/auth/signin" element={<Signin onSignIn={onSignIn} />} />
+            <Route path="/auth/signup" element={<Signup onSignIn={onSignIn} />} />
+          </Routes>
         </Router>
-      </StylesProvider>
+      </StyledEngineProvider>
     </div>
   );
 };
