@@ -1,10 +1,10 @@
-import React from "react";
-import ReactDOM from "react-dom";
-import { createMemoryHistory, createBrowserHistory } from "history";
-import App from "./App";
+import React from 'react';
+import { createRoot } from 'react-dom/client';
+import { createMemoryHistory, createBrowserHistory } from 'history';
+import App from './App';
 
 // Mount function to start up the app
-const mount = (el, { onNavigate, defaultHistory, initialPath }) => {
+const mount = (el, { onSignIn, onNavigate, defaultHistory, initialPath }) => {
   const history =
     defaultHistory ||
     createMemoryHistory({
@@ -12,10 +12,13 @@ const mount = (el, { onNavigate, defaultHistory, initialPath }) => {
     });
 
   if (onNavigate) {
-    history.listen(onNavigate);
+    history.listen((update) => {
+      onNavigate({ pathname: update.location.pathname });
+    });
   }
 
-  ReactDOM.render(<App history={history} />, el);
+  const root = createRoot(el);
+  root.render(<App onSignIn={onSignIn} history={history} />);
 
   return {
     onParentNavigate({ pathname: nextPathname }) {
@@ -30,8 +33,8 @@ const mount = (el, { onNavigate, defaultHistory, initialPath }) => {
 
 // If we are in development and in isolation,
 // call mount immediately
-if (process.env.NODE_ENV === "development") {
-  const devRoot = document.querySelector("#_auth-dev-root");
+if (process.env.NODE_ENV === 'development') {
+  const devRoot = document.querySelector('#_auth-dev-root');
 
   if (devRoot) {
     mount(devRoot, { defaultHistory: createBrowserHistory() });
