@@ -1,28 +1,32 @@
-import React from 'react';
-import { Switch, Route, Router } from 'react-router-dom';
-import {
-  StylesProvider,
-  createGenerateClassName,
-} from '@material-ui/core/styles';
+import React, { useState, useEffect } from 'react';
+import { Routes, Route, Router } from 'react-router-dom';
+import { StyledEngineProvider } from '@mui/material/styles';
 
 import Landing from './components/Landing';
 import Pricing from './components/Pricing';
 
-const generateClassName = createGenerateClassName({
-  productionPrefix: 'ma',
-});
-
 export default ({ history }) => {
+  const [location, setLocation] = useState(history.location);
+
+  useEffect(() => {
+    // Listen to history changes and update location state
+    const unlisten = history.listen((update) => {
+      setLocation(update.location);
+    });
+
+    return unlisten; // Clean up the listener on unmount
+  }, [history]);
+
   return (
     <div>
-      <StylesProvider generateClassName={generateClassName}>
-        <Router history={history}>
-          <Switch>
-            <Route exact path="/pricing" component={Pricing} />
-            <Route path="/" component={Landing} />
-          </Switch>
+      <StyledEngineProvider injectFirst>
+        <Router location={location} navigator={history}>
+          <Routes>
+            <Route path="/pricing" element={<Pricing />} />
+            <Route path="/" element={<Landing />} />
+          </Routes>
         </Router>
-      </StylesProvider>
+      </StyledEngineProvider>
     </div>
   );
 };
