@@ -1,21 +1,21 @@
 import React from 'react';
-import Button from '@material-ui/core/Button';
-import Card from '@material-ui/core/Card';
-import CardActions from '@material-ui/core/CardActions';
-import CardContent from '@material-ui/core/CardContent';
-import CardMedia from '@material-ui/core/CardMedia';
-import Grid from '@material-ui/core/Grid';
-import Typography from '@material-ui/core/Typography';
-import { makeStyles } from '@material-ui/core/styles';
-import Container from '@material-ui/core/Container';
-import MaterialLink from '@material-ui/core/Link';
+import Button from '@mui/material/Button';
+import Card from '@mui/material/Card';
+import CardActions from '@mui/material/CardActions';
+import CardContent from '@mui/material/CardContent';
+import CardMedia from '@mui/material/CardMedia';
+import Grid from '@mui/material/Grid';
+import Typography from '@mui/material/Typography';
+import Container from '@mui/material/Container';
+import MaterialLink from '@mui/material/Link';
 import { Link } from 'react-router-dom';
+import { Box } from '@mui/material';
 
 function Copyright() {
   return (
     <Typography variant="body2" color="textSecondary" align="center">
       {'Copyright © '}
-      <MaterialLink component={Link} to="/" color="inherit">
+      <MaterialLink component={Link} to="/" color="inherit" sx={{ textDecoration: 'none' }}>
         Your Website
       </MaterialLink>{' '}
       {new Date().getFullYear()}
@@ -24,53 +24,20 @@ function Copyright() {
   );
 }
 
-const useStyles = makeStyles((theme) => ({
-  '@global': {
-    a: {
-      textDecoration: 'none',
-    },
-  },
-  icon: {
-    marginRight: theme.spacing(2),
-  },
-  heroContent: {
-    backgroundColor: theme.palette.background.paper,
-    padding: theme.spacing(8, 0, 6),
-  },
-  heroButtons: {
-    marginTop: theme.spacing(4),
-  },
-  cardGrid: {
-    paddingTop: theme.spacing(8),
-    paddingBottom: theme.spacing(8),
-  },
-  card: {
-    height: '100%',
-    display: 'flex',
-    flexDirection: 'column',
-  },
-  cardMedia: {
-    paddingTop: '56.25%', // 16:9
-  },
-  cardContent: {
-    flexGrow: 1,
-  },
-  footer: {
-    backgroundColor: theme.palette.background.paper,
-    padding: theme.spacing(6),
-  },
-}));
-
 const cards = [1, 2, 3, 4, 5, 6, 7, 8, 9];
 
 export default function Album() {
-  const classes = useStyles();
-
   return (
     <React.Fragment>
       <main>
         {/* Hero unit */}
-        <div className={classes.heroContent}>
+        <Box
+          sx={{
+            backgroundColor: 'background.paper',
+            pt: 8,
+            pb: 6,
+          }}
+        >
           <Container maxWidth="sm">
             <Typography
               component="h1"
@@ -91,38 +58,46 @@ export default function Album() {
               contents, the creator, etc. Make it short and sweet, but not too
               short so folks don&apos;t simply skip over it entirely.
             </Typography>
-            <div className={classes.heroButtons}>
+            <Box sx={{ mt: 4 }}>
               <Grid container spacing={2} justifyContent="center">
                 <Grid item>
-                  <Link to="/pricing">
+                  <Link to="/pricing" style={{ textDecoration: 'none' }}>
                     <Button variant="contained" color="primary">
                       Pricing
                     </Button>
                   </Link>
                 </Grid>
                 <Grid item>
-                  <Link to="/pricing">
+                  <Link to="/pricing" style={{ textDecoration: 'none' }}>
                     <Button variant="outlined" color="primary">
                       Pricing
                     </Button>
                   </Link>
                 </Grid>
               </Grid>
-            </div>
+            </Box>
           </Container>
-        </div>
-        <Container className={classes.cardGrid} maxWidth="md">
+        </Box>
+        <Container sx={{ py: 8 }} maxWidth="md">
           {/* End hero unit */}
           <Grid container spacing={4}>
             {cards.map((card) => (
               <Grid item key={card} xs={12} sm={6} md={4}>
-                <Card className={classes.card}>
+                <Card
+                  sx={{
+                    height: '100%',
+                    display: 'flex',
+                    flexDirection: 'column',
+                  }}
+                >
                   <CardMedia
-                    className={classes.cardMedia}
-                    image="https://source.unsplash.com/random"
+                    sx={{
+                      pt: '56.25%', // 16:9
+                    }}
+                    image={`https://picsum.photos/400/300?random=${card}`}
                     title="Image title"
                   />
-                  <CardContent className={classes.cardContent}>
+                  <CardContent sx={{ flexGrow: 1 }}>
                     <Typography gutterBottom variant="h5" component="h2">
                       Heading
                     </Typography>
@@ -146,7 +121,13 @@ export default function Album() {
         </Container>
       </main>
       {/* Footer */}
-      <footer className={classes.footer}>
+      <Box
+        component="footer"
+        sx={{
+          backgroundColor: 'background.paper',
+          py: 6,
+        }}
+      >
         <Typography variant="h6" align="center" gutterBottom>
           Footer
         </Typography>
@@ -159,7 +140,7 @@ export default function Album() {
           Something here to give the footer a purpose!
         </Typography>
         <Copyright />
-      </footer>
+      </Box>
       {/* End footer */}
     </React.Fragment>
   );
