@@ -1,13 +1,13 @@
-const { merge } = require("webpack-merge");
-const HtmlWebpackPlugin = require("html-webpack-plugin");
-const ModuleFederationPlugin = require("webpack/lib/container/ModuleFederationPlugin");
-const commonConfig = require("./webpack.common");
-const packageJson = require("../package.json");
+const { merge } = require('webpack-merge');
+const HtmlWebpackPlugin = require('html-webpack-plugin');
+const ModuleFederationPlugin = require('webpack/lib/container/ModuleFederationPlugin');
+const commonConfig = require('./webpack.common');
+const packageJson = require('../package.json');
 
 const devConfig = {
-  mode: "development",
+  mode: 'development',
   output: {
-    publicPath: "http://localhost:8082/",
+    publicPath: 'http://localhost:8082/',
   },
   devServer: {
     port: 8082,
@@ -17,15 +17,29 @@ const devConfig = {
   },
   plugins: [
     new ModuleFederationPlugin({
-      name: "auth",
-      filename: "remoteEntry.js",
+      name: 'auth',
+      filename: 'remoteEntry.js',
       exposes: {
-        "./AuthApp": "./src/bootstrap",
+        './AuthApp': './src/bootstrap',
       },
-      shared: packageJson.dependencies,
+      shared: {
+        ...packageJson.dependencies,
+        react: {
+          singleton: true,
+          requiredVersion: packageJson.dependencies.react,
+        },
+        'react-dom': {
+          singleton: true,
+          requiredVersion: packageJson.dependencies['react-dom'],
+        },
+        'react-router-dom': {
+          singleton: true,
+          requiredVersion: packageJson.dependencies['react-router-dom'],
+        },
+      },
     }),
     new HtmlWebpackPlugin({
-      template: "./public/index.html",
+      template: './public/index.html',
     }),
   ],
 };
