@@ -7,36 +7,25 @@ const packageJson = require('../package.json');
 const devConfig = {
   mode: 'development',
   output: {
-    publicPath: 'http://localhost:8081/',
+    publicPath: 'http://localhost:8083/',
   },
   devServer: {
-    port: 8081,
+    port: 8083,
     historyApiFallback: {
       historyApiFallback: true,
+    },
+    headers: {
+      'Access-Control-Allow-Origin': '*',
     },
   },
   plugins: [
     new ModuleFederationPlugin({
-      name: 'marketing',
+      name: 'dashboard',
       filename: 'remoteEntry.js',
       exposes: {
-        './MarketingApp': './src/bootstrap',
+        './DashboardApp': './src/bootstrap',
       },
-      shared: {
-        ...packageJson.dependencies,
-        react: {
-          singleton: true,
-          requiredVersion: packageJson.dependencies.react,
-        },
-        'react-dom': {
-          singleton: true,
-          requiredVersion: packageJson.dependencies['react-dom'],
-        },
-        'react-router-dom': {
-          singleton: true,
-          requiredVersion: packageJson.dependencies['react-router-dom'],
-        },
-      },
+      shared: packageJson.dependencies,
     }),
     new HtmlWebpackPlugin({
       template: './public/index.html',
